+++ easy-ui-react/src/TabNav/TabNav.mdx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Canvas, Meta, ArgTypes, Controls } from "@storybook/blocks";
+import { TabNav } from "./TabNav";
+import * as TabNavStories from "./TabNav.stories";
+
+<Meta of={TabNavStories} />
+
+# TabNav
+
+A `TabNav` is a set of styled links that let users navigate between related pages.
+
+<Canvas of={TabNavStories.Default} />
