+++ easy-ui-react/src/TabNav/TabNav.module.scss
@@ -34,22 +34,3 @@
   transition-duration: 0.2s;
   transition-timing-function: ease-in-out;
 }
-
-.TabNavItem {
-  @include unstyled.button;
-
-  display: inline-flex;
-  padding: 0 design-token("space.0.5");
-  margin: 0 -4px -1px -4px;
-  color: hsla(227, 83%, 14%, 1);
-  cursor: pointer;
-  text-decoration: none;
-}
-
-.hovered {
-  color: hsla(226, 84%, 27%, 1);
-}
-
-.selected {
-  color: hsla(226, 100%, 54%, 1);
-}
