+++ easy-ui-react/src/TabNav/TabNav.tsx
@@ -1,18 +1,8 @@
 import { AriaLabelingProps } from "@react-types/shared";
-import React, {
-  ComponentProps,
-  ElementType,
-  ReactNode,
-  createContext,
-  useContext,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
-import { useHover } from "react-aria";
-import { Text } from "../Text";
-import { classNames, getComponentToken } from "../utilities/css";
+import React, { ReactNode, useMemo, useState } from "react";
+import { getComponentToken } from "../utilities/css";
+import { TabNavItem } from "./TabNavItem";
+import { TabNavContext } from "./context";
 
 import styles from "./TabNav.module.scss";
 
@@ -20,36 +10,21 @@ type TabNavProps = AriaLabelingProps & {
   children: ReactNode;
 };
 
-type TabNavItemProps<T extends ElementType = "a"> = ComponentProps<T> & {
-  as?: T;
-  children: ReactNode;
-  isCurrentPage?: boolean;
-};
-
-const TabNavContext = createContext<{
-  setLeft: React.Dispatch<React.SetStateAction<number | null>>;
-  setWidth: React.Dispatch<React.SetStateAction<number | null>>;
-} | null>(null);
-
-const useTabNav = () => {
-  const tabNavContext = useContext(TabNavContext);
-  if (!tabNavContext) {
-    throw new Error("useTabNav must be used within a TabNav");
-  }
-  return tabNavContext;
-};
-
 export function TabNav(props: TabNavProps) {
   const { children, ...labelingProps } = props;
+
   const [width, setWidth] = useState<number | null>(null);
   const [left, setLeft] = useState<number | null>(null);
+
   const context = useMemo(() => {
     return { setWidth, setLeft };
   }, []);
+
   const style = {
     ...getComponentToken("tab-nav", "selected-tab-width", `${width}px`),
     ...getComponentToken("tab-nav", "selected-tab-left", `${left}px`),
   };
+
   return (
     <TabNavContext.Provider value={context}>
       <nav {...labelingProps} className={styles.TabNav} style={style}>
@@ -59,38 +34,6 @@ export function TabNav(props: TabNavProps) {
         </ul>
       </nav>
     </TabNavContext.Provider>
-  );
-}
-
-function TabNavItem<T extends ElementType = "a">(props: TabNavItemProps<T>) {
-  const { setWidth, setLeft } = useTabNav();
-  const { as: As = "a", children, isCurrentPage, ...restProps } = props;
-  const ref = useRef<HTMLLIElement | null>(null);
-  const { isHovered, hoverProps } = useHover({});
-  const className = classNames(
-    styles.TabNavItem,
-    isCurrentPage && styles.selected,
-    isHovered && styles.hovered,
-  );
-  useEffect(() => {
-    if (isCurrentPage && ref.current && ref.current.parentElement) {
-      const parentRect = ref.current.parentElement.getBoundingClientRect();
-      const rect = ref.current.getBoundingClientRect();
-      setWidth(rect.width);
-      setLeft(rect.x - parentRect.x);
-    }
-  }, [setWidth, setLeft, isCurrentPage]);
-  return (
-    <li ref={ref}>
-      <As
-        {...hoverProps}
-        className={className}
-        aria-current={isCurrentPage ? "page" : undefined}
-        {...restProps}
-      >
-        <Text variant="subtitle1">{children}</Text>
-      </As>
-    </li>
   );
 }
