+++ easy-ui-react/src/TabNav/TabNavItem.module.scss
@@ -1,0 +1,23 @@
+@use "../styles/common" as *;
+@use "../styles/unstyled";
+
+.TabNavItem {
+  @include unstyled.button;
+  @include font-style("subtitle1");
+
+  display: inline-flex;
+  padding: 0 design-token("space.0.5");
+  margin: 0 -4px -1px -4px;
+
+  color: hsla(227, 83%, 14%, 1);
+  cursor: pointer;
+  text-decoration: none;
+}
+
+.hovered {
+  color: hsla(226, 84%, 27%, 1);
+}
+
+.selected {
+  color: hsla(226, 100%, 54%, 1);
+}
